Grand Clearance: Delete all files to restart work
It's been a long time since my last update to repo,
and there'd been unfathomably many changes in Android.
This project is not abandoned but shall be restarted
from the baseline of concept.
Things will start from the establishing objective, design,
user interface and all functionalities.

It takes time, but will arrive finally.